+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-# next
+# 3.0.0-beta.11
 
 - `positionsItselfAndKeepDistance` little improvements
 - `bonfireCamera.shake` improvements